+++ changelog.md
@@ -4,6 +4,7 @@
 - get(), getOnly() function now accept closure for default value.
 - set() function now accepts closure for value.
 - Arrayable interface from [Illuminate\Contracts](https://github.com/illuminate/contracts) added to AbstractDict.
+- dot operation support added to call() function.
 
 ## v1.0.0
 - Old versions deleted.